A pair of functions that cache the inverse of a matrix
The first function  creates a special "matrix" object that can cache its
inverse.
Second function computes the inverse of the special "matrix" returned by
first function. If the inverse has already been calculated (and the
matrix has not changed), then it retrieve the inverse from the cache.
otherwise it will calculate and return the inverse.