functions that cache the inverse of a matrix
First function creates a special "matrix" object, which is a list
containing function to
## 1. set the value of the matrix
## 2. get the value of the matrix
## 3. set the inverse of the matrix
## 4. get the cached inverse of the matrix

Second function computes the inverse of the special "matrix" returned by
function above.
If the inverse has already been calculated (and the matrix has not
changed),  then function will retrieve the inverse from the cache.
Otherwise, it will calculate and return the inverse of the matrix.